fix: e2e webpack compile error.
Make new CctConstants.ts file (rather than the .tsx file that causes issue with e2e test webpack compilation)
to hold the cctCalcWarningsMsgs.
fix: the targeting of the saved cct calc from list (as we can't rely on the number prefix).

Testing is 💩
