+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "trainee-ui-app",
-  "version": "0.109.0",
+  "version": "0.109.1",
   "private": true,
   "dependencies": {
     "@aws-amplify/ui-react": "^5.3.0",
