+++ utilities/Constants.tsx
@@ -553,15 +553,3 @@ export const fteOptions = [
   { value: 60, label: "60%" },
   { value: 50, label: "50%" }
 ];
-
-// cct calc warning messages
-export const cctCalcWarningsMsgs = {
-  noActiveProgsMsg:
-    "You do not have any current, upcoming or future programmes to link to a CCT calculation.",
-  shortNoticeMsg:
-    "Changing your WTE hours with less than 16 week's notice may not be possible (e.g. rota constraints).",
-  wteIncreaseMsg:
-    "Increasing your WTE hours requires a suitable post to be available, which may not be possible.",
-  wteCustomMsg:
-    "A custom WTE change may not be possible; it will require Dean approval."
-};
